[feat] : figma를 통해 css 기본 속성 정의
