+++ css/variables.css
@@ -1,0 +1,61 @@
+.title {
+  font-family: Roboto;
+  font-weight: 800;
+  font-size: 24px;
+  line-height: 36px;
+  letter-spacing: 0px;
+}
+
+.subtitle {
+  font-family: Roboto;
+  font-weight: 600;
+  font-size: 20px;
+  line-height: 24px;
+  letter-spacing: 0.15px;
+}
+
+.body {
+  font-family: Roboto;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 24px;
+  letter-spacing: 0.5px;
+}
+
+.placeholder {
+  font-family: Roboto;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 24px;
+  letter-spacing: 0.5px;
+}
+
+.button,
+.catption {
+  width: 180;
+  height: 24;
+}
+
+.lotto-primary {
+  background: #4e5ba6;
+}
+
+.lotto-greyscale-1 {
+  background: #ffffff;
+}
+
+.lotto-greyscale-2 {
+  background: #fcfcfd;
+}
+
+.lotto-greyscale-3 {
+  background: #b4b4b4;
+}
+
+.lotto-greyscale-4 {
+  background: #8b8b8b;
+}
+
+.lotto-greyscale-5 {
+  background: #000000;
+}
