+++ src/utils/validate/Validator.js
@@ -1,3 +1,4 @@
+import { transformWithEsbuild } from "vite";
 import Validate from "./Validate.js";
 
 const validate = new Validate();
